refactor(indexer/base): move to cosmossdk.io/schema, add ApplyUpdate callback
