+++ schema/go.mod
@@ -1,4 +1,5 @@
-module cosmossdk.io/indexer/base
+module cosmossdk.io/schema
+
 
 // NOTE: this go.mod should have zero dependencies and remain on go 1.12 to stay compatible
 // with all known production releases of the Cosmos SDK. This is to ensure that all historical
